Refactor estructura de carpetas
