+++ .idea/androidTestResultsUserPreferences.xml
@@ -153,6 +153,7 @@
                 <map>
                   <entry key="Duration" value="90" />
                   <entry key="Pixel_3_API_VanillaIceCream" value="120" />
+                  <entry key="Pixel_8_Pro_API_VanillaIceCream" value="120" />
                   <entry key="Tests" value="360" />
                 </map>
               </option>
