+++ app/src/androidTest/java/com/miso/appvinilos/pruebasunitarias/DetalleAlbumTest.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.test.performScrollTo
 import androidx.compose.ui.test.performScrollToIndex
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.compose.rememberNavController
-import com.miso.appvinilos.albums.model.Album
+import com.miso.appvinilos.data.model.Album
 import com.miso.appvinilos.albums.ui.views.AlbumBasicDetail
 import com.miso.appvinilos.albums.ui.views.AlbumList
 import com.miso.appvinilos.albums.viewmodels.AlbumViewModel
@@ -23,7 +23,7 @@ import org.junit.Rule
 
 import org.junit.Assert.*
 class detalleAlbumTest {
-    val albumTestNormal=Album(id = 1, name="Album1", cover="cover1", releaseDate="2021-01-01", description="description1", genre="genre1", recordLabel="recordLabel1")
+    val albumTestNormal= Album(id = 1, name="Album1", cover="cover1", releaseDate="2021-01-01", description="description1", genre="genre1", recordLabel="recordLabel1")
 
     @get:Rule
     val composeTestRule = createAndroidComposeRule<ComponentActivity>()
