+++ app/src/main/java/com/miso/appvinilos/data/repositories/AlbumRepository.kt
@@ -1,6 +1,6 @@
-package com.miso.appvinilos.albums.repositories
-import com.miso.appvinilos.albums.model.Album
-import com.miso.appvinilos.albums.network.AlbumsApi
+package com.miso.appvinilos.data.repositories
+import com.miso.appvinilos.data.model.Album
+import com.miso.appvinilos.data.network.AlbumsApi
 
 class AlbumRepository {
     private val albumService = AlbumsApi.albumsService
